requirements.txt
google-cloud-storage==2.11.0
pandas==2.0.2
requests==2.31.0
beautifulsoup4==4.12.2
chardet==5.1.0
openpyxl==3.1.2
